refactoring on package naming
