+++ src/main/java/com/gym/dao/AddressJpaDao.java
@@ -1,0 +1,9 @@
+package com.gym.dao;
+
+import com.gym.entity.Address;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AddressJpaDao extends JpaRepository<Address, String> {
+}
