+++ src/main/java/com/gym/dao/GroupTrainingJpaDao.java
@@ -1,0 +1,9 @@
+package com.gym.dao;
+
+import com.gym.entity.GroupTraining;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GroupTrainingJpaDao extends JpaRepository<GroupTraining, String> {
+}
