+++ src/main/java/com/gym/entity/Customer.java
@@ -1,0 +1,140 @@
+package com.gym.entity;
+
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+@Entity
+@Table(name = "CUSTOMERS")
+public class Customer implements Serializable {
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
+    @Column(name = "ID")
+    private String id;
+    @Column(name = "PERSON_NUMBER")
+    private String personNumber;
+    @Column(name = "FIRST_NAME")
+    private String firstName;
+    @Column(name = "MIDDLE_NAME")
+    private String middleName;
+    @Column(name = "LAST_NAME")
+    private String lastName;
+    @Column(name = "PHONE_NUMBER")
+    private String phoneNumber;
+    @OneToOne(mappedBy = "customer", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
+    private Address address;
+    @OneToOne(mappedBy = "customer", cascade = CascadeType.PERSIST, orphanRemoval = true, fetch = FetchType.LAZY)
+    private Subscription subscription;
+    @Column(name = "NUMBER_OF_BOOKINGS_PER_WEEK")
+    private int numberOfBookingAllowedPerWeek;
+    @Column(name = "MEMBER_SINCE")
+    private String memberSince;
+
+    public Customer() {
+    }
+
+    public Customer(String id, String personNumber, String firstName, String middleName, String lastName, String phoneNumber, Address address, Subscription subscription, int numberOfBookingAllowedPerWeek, String memberSince) {
+        this.id = id;
+        this.personNumber = personNumber;
+        this.firstName = firstName;
+        this.middleName = middleName;
+        this.lastName = lastName;
+        this.phoneNumber = phoneNumber;
+        this.address = address;
+        this.subscription = subscription;
+        this.numberOfBookingAllowedPerWeek = numberOfBookingAllowedPerWeek;
+        this.memberSince = memberSince;
+    }
+
+    public Customer(String personNumber, String firstName, String middleName, String lastName, String phoneNumber, int numberOfBookingAllowedPerWeek, String memberSince) {
+        this.personNumber = personNumber;
+        this.firstName = firstName;
+        this.middleName = middleName;
+        this.lastName = lastName;
+        this.phoneNumber = phoneNumber;
+        this.numberOfBookingAllowedPerWeek = numberOfBookingAllowedPerWeek;
+        this.memberSince = memberSince;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getPersonNumber() {
+        return personNumber;
+    }
+
+    public void setPersonNumber(String personNumber) {
+        this.personNumber = personNumber;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getMiddleName() {
+        return middleName;
+    }
+
+    public void setMiddleName(String middleName) {
+        this.middleName = middleName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public Address getAddress() {
+        return address;
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+
+    public Subscription getMembership() {
+        return subscription;
+    }
+
+    public void setMembership(Subscription subscription) {
+        this.subscription = subscription;
+    }
+
+    public int getNumberOfBookingAllowedPerWeek() {
+        return numberOfBookingAllowedPerWeek;
+    }
+
+    public void setNumberOfBookingAllowedPerWeek(int numberOfBookingAllowedPerWeek) {
+        this.numberOfBookingAllowedPerWeek = numberOfBookingAllowedPerWeek;
+    }
+
+    public String getMemberSince() {
+        return memberSince;
+    }
+
+    public void setMemberSince(String memberSince) {
+        this.memberSince = memberSince;
+    }
+}
