+++ src/main/java/com/gym/exception/CustomerNotFoundException.java
@@ -1,0 +1,5 @@
+package com.gym.exception;
+
+public class CustomerNotFoundException extends Exception {
+    public CustomerNotFoundException(String msg){super(msg);}
+}
