+++ src/main/java/com/gym/service/SubscriptionService.java
@@ -1,0 +1,16 @@
+package com.gym.service;
+
+
+import com.gym.dto.SubscriptionDto;
+import com.gym.entity.Subscription;
+import com.gym.exception.CustomerNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface SubscriptionService {
+    Subscription AddSubscription(String customerId, SubscriptionDto subscriptionDto) throws CustomerNotFoundException;
+
+    Subscription updateSubscription(String customerId, SubscriptionDto subscriptionDto) throws CustomerNotFoundException;
+
+    Subscription getSubscription(String customerId) throws CustomerNotFoundException;
+}
