--- src/main/java/develop/gym/GymApplication.java
@@ -1,12 +1,0 @@
-package develop.gym;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GymApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(GymApplication.class, args);
-    }
-
-}
