--- src/main/java/develop/gym/dao/AddressJpaDao.java
@@ -1,9 +1,0 @@
-package develop.gym.dao;
-
-import develop.gym.entity.Address;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AddressJpaDao extends JpaRepository<Address, String> {
-}
