--- src/main/java/develop/gym/exception/CustomerNotFoundException.java
@@ -1,5 +1,0 @@
-package develop.gym.exception;
-
-public class CustomerNotFoundException extends Exception {
-    public CustomerNotFoundException(String msg){super(msg);}
-}
