--- src/main/java/develop/gym/service/AddressService.java
@@ -1,15 +1,0 @@
-package develop.gym.service;
-
-import develop.gym.dto.AddressDto;
-import develop.gym.entity.Address;
-import develop.gym.exception.CustomerNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public interface AddressService {
-    Address saveAddress(String customerId, AddressDto addressDto) throws CustomerNotFoundException;
-
-    Address updateAddress(String customerId, AddressDto addressDto) throws CustomerNotFoundException;
-
-    Address getAddressForCustomer(String customerId) throws CustomerNotFoundException;
-}
